backend/requirements.txt
# DataGenesis AI Backend Dependencies
# Updated to resolve dependency conflicts

# Core FastAPI and server
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.5.0,<3.0.0

# Database and storage
supabase>=2.0.0,<3.0.0
redis>=5.0.0,<6.0.0

# HTTP client - let other packages determine compatible version
httpx>=0.24.0

# Authentication and security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.0,<2.0.0

# File handling
python-multipart>=0.0.6,<1.0.0
aiofiles>=23.2.0,<24.0.0

# AI and ML services
google-generativeai>=0.3.0,<1.0.0
google-cloud-aiplatform>=1.36.0,<2.0.0

# Data processing
pandas>=2.1.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0

# Vector database (optional)
pinecone-client>=3.0.0,<4.0.0

# Background tasks
celery>=5.3.0,<6.0.0

# WebSockets
websockets>=12.0,<13.0

# Environment management
python-dotenv>=1.0.0,<2.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0

# Additional utilities for production
gunicorn>=21.2.0,<22.0.0
python-json-logger>=2.0.0,<3.0.0